Ugly version of event console
For users with permissions to perform certain actions, users will be
presented with the buttons that can perform those actions (permissions
can be set in the django admin, preferably using groups.) For actions
that can be done to other people, particularly officers signing other
people out, a table is presented with users in the primary column and
actions in other columns.

I expect to add other data to the table, to allow officers to view more
details about users in the table. Also, I undid the error notifications;
I will add them back (hopefully better) eventually.